+++ Assets/Scripts/2D/Projectiles/ExplodeBullet2D.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ExplodeBullet2D : Projectile2D
 {
-    public void SetData(float damage, float critChance, Quaternion rotation, Vector3 direction, float speed, Vector3 pos, string axis, Explode2D explode)
+    public void SetData(int damage, float critChance, Quaternion rotation, Vector3 direction, float speed, Vector3 pos, string axis, Explode2D explode)
     {
         SetData(damage, critChance, rotation, direction, speed, pos, axis);
         this.explode = explode;
